requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
Pillow==5.3.0
requests==2.20.1
urllib3==1.24.1
Werkzeug==0.14.1
